Real bevezetese+++

+++ Alkatrész.cpp
@@ -23,7 +23,7 @@ void MainWindow::alk_krv_rajzol(int alk)
 	for (int k = 0; k <= 1; k++)
 	{
 		if (a.v.size() > 0)
-			for (int i = 0; i < a.v.size(); i++)
+			for (size_t i = 0; i < a.v.size(); i++)
 				if (a.v[i].szint == sz[k])
 				{
 					VONAL v = a.v[i];
